jenkins_k8s/generator/requirements.txt
jenkins
gitPython
